requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.8.2
SQLAlchemy==2.0.32
alembic==1.13.2
python-dotenv==1.0.1
pyteal==0.26.1
py-algorand-sdk==2.6.1
